Move StateInternals to runners/core-java
