+++ runners/core-java/src/main/java/org/apache/beam/runners/core/StateInternalsFactory.java
@@ -20,7 +20,6 @@ package org.apache.beam.runners.core;
 import java.io.Serializable;
 import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
-import org.apache.beam.sdk.util.state.StateInternals;
 
 /**
  * A factory for providing {@link StateInternals} for a particular key.
